requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
configparser==5.0.2
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
